Add new options to clap flags in options.rs
New command line options, 'pe_chunk' and 'att_chunk', have been added to the options.rs file. These modifications improve the flexibility and customizability of the program by allowing the user to input the values for these two new options. The user's experience and the program's functionality have been enhanced with this update.